Main layout columns are now grid column, instead of another stack panel
This ensures that the contents of each grid column have a well-defined
size; a stack panel will impose no size on the child it's laying out, so
it doesn't get much information. With these columns, the first column
will size itself to the child, but then the main content area (the map
itself) will be given the remaining amount of space for whatever wants
to go there.