+Custom panel that renders the images outside of the layout system
This is a quick and dirty swing at a different approach. Laying out
19000 tiles in the WPF layout system is real bad, and uses way more than
we need. These are ultimately just images in a grid, so we can use that
specificity to reduce the complexity a ton.

This panel binds a collection of tile view models and iterates them and
renders their images. This makes some assumptions, such as columns and
rows, and ideally wouldn't know what a view model was. It also can't do
scrolling or user interaction. But it's hella fast.