+Sprite class
This abstraction is designed to hold image information in a format
useful for WPF. This class can load PNGs from disk and convert their
indexed palette color information to 32-bit colors, which is much easier
to work with for UIs.

This presents a GetPixel/SetPixel API with Color instances, making it
really easy to work with, exempting the caller from reasoning through
buffer offsets.

As proof that it can read files, I added a SaveAsPng method to show that
ingested paths can be exported as 32-bit PNGs.