Clicking a tile on the map changes it to forest
This fleshes out the ability know when a tile is clicked, despite
pan/zoom values. It also alters the cell's index in the associated view
model, then updates the "single bitmap," and redraws the image. It's
hard-coded to use the forest tile, but choosing which tile to draw with
is up next.