Can now select tiles and then click the map to change to that tile
This uses a binding to the tile view model's Index property. I made this
now have a public getter, rather than having the SetIndex() method. This
allows the binding to work, and presents a more common symmetrical
get/set API. The extra side-effects and state reevaluation are now part
of the setter. Also note that the initial value is now -1, to ensure
that the Set occurs; when defaulting to 0, and then assigning to 0, the
set doesn't actually occur since the value didn't change, and this
didn't assign the image.

I also removed the custom template for the list box items just to get
the basics going, even though it looks a little crappy. The selection
state is clear, though.